requirements.txt
streamlit==1.22.0
boto3==1.26.135
Pillow==9.5.0